# public/swagger-specs/services/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ConsenSys Codefi Orchestrate API. Enables dynamic management of transactions, identities, chains, faucets and contracts.\nTransaction Requests are an abstraction over schedules and jobs representing one or more transactions executed on the Blockchain network\nSchedules are ordered lists of jobs executed in a predefined sequence\nJobs represent single transaction flows executed on the Blockchain network\nChains represent list of endpoints pointing to a Blockchain network\nFaucets represent funded accounts (holding ETH) linked to specific chains, allowed to fund newly created accounts automatically for them to be able to send transactions.\nAccounts represent Ethereum accounts (private keys). By usage of the generated cryptographic key pair, accounts can be used to sign/verify and to encrypt/decrypt messages.\nContracts represent Solidity contracts management.\nEvent Streams represent Event streams management.",
        "title": "Orchestrate API",
        "contact": {
            "name": "Contact ConsenSys Codefi Orchestrate",
            "url": "https://consensys.net/codefi/orchestrate/contact",
            "email": "orchestrate@consensys.net"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "paths": {
...
        "/eventstreams": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Get a list of filtered event streams",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event Streams"
                ],
                "summary": "Search event streams by provided filters",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "List of event stream names",
                        "name": "names",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Chain UUID",
                        "name": "chain_uuid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tenant ID",
                        "name": "tenant_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of event streams found",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.EventStreamResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid filter in the request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/eventstreams/kafka": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Creates a new Kafka Event stream",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event Streams"
                ],
                "summary": "Creates a new Kafka Event stream",
                "parameters": [
                    {
                        "description": "Kafka Event stream creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateKafkaEventStreamRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event stream object",
                        "schema": {
                            "$ref": "#/definitions/types.EventStreamResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/eventstreams/kafka/{uuid}": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Update a specific Kafka event stream by uuid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event Streams"
                ],
                "summary": "Update Kafka event stream by uuid",
                "parameters": [
                    {
                        "description": "Kafka Event stream update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateKafkaEventStreamRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "event stream uuid",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event stream found",
                        "schema": {
                            "$ref": "#/definitions/types.EventStreamResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Account not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/eventstreams/webhooks": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Creates a new Webhook Event stream",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event Streams"
                ],
                "summary": "Creates a new Webhook Event stream",
                "parameters": [
                    {
                        "description": "Webhook Event stream creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateWebhookEventStreamRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event stream object",
                        "schema": {
                            "$ref": "#/definitions/types.EventStreamResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/eventstreams/webhooks/{uuid}": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Update a specific Webhook event stream by uuid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event Streams"
                ],
                "summary": "Update Webhook event stream by uuid",
                "parameters": [
                    {
                        "description": "Webhook Event stream update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateWebhookEventStreamRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "event stream uuid",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event stream found",
                        "schema": {
                            "$ref": "#/definitions/types.EventStreamResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Account not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/eventstreams/{uuid}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Fetch a single event stream by uuid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event Streams"
                ],
                "summary": "Fetch an event stream by uuid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "event stream uuid",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event stream found",
                        "schema": {
                            "$ref": "#/definitions/types.EventStreamResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Event stream not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event Streams"
                ],
                "summary": "Deletes an event stream by uuid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid of the event stream",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Event Stream not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 24000
                },
                "message": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "types.CreateKafkaEventStreamRequest": {
            "type": "object",
            "required": [
                "name",
                "topic"
            ],
            "properties": {
                "chain": {
                    "type": "string",
                    "example": "mainnet"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "my-kafka-stream"
                },
                "topic": {
                    "type": "string",
                    "example": "my-notification-topic"
                }
            }
        },
...
        "types.CreateWebhookEventStreamRequest": {
            "type": "object",
            "required": [
                "name",
                "url"
            ],
            "properties": {
                "chain": {
                    "type": "string",
                    "example": "mainnet"
                },
                "headers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "my-webhook-stream"
                },
                "url": {
                    "type": "string",
                    "example": "https://my-event-steam-endpoint.com"
                }
            }
        },
...
        "types.EventStreamResponse": {
            "type": "object",
            "properties": {
                "chain_uuid": {
                    "type": "string",
                    "example": "b4374e6f-b28a-4bad-b4fe-bda36eaf849c"
                },
                "channel": {
                    "type": "string",
                    "example": "webhook"
                },
                "createdAt": {
                    "type": "string",
                    "example": "2020-07-09T12:35:42.115395Z"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "my-stream"
                },
                "ownerID": {
                    "type": "string",
                    "example": "foo"
                },
                "specs": {
                    "type": "object"
                },
                "status": {
                    "type": "string",
                    "example": "LIVE"
                },
                "tenantID": {
                    "type": "string",
                    "example": "foo"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2020-07-09T12:35:42.115395Z"
                },
                "uuid": {
                    "type": "string",
                    "example": "b4374e6f-b28a-4bad-b4fe-bda36eaf849c"
                }
            }
        },
...
        "types.UpdateKafkaEventStreamRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "chain": {
                    "type": "string",
                    "example": "mainnet"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "my-kafka-stream"
                },
                "status": {
                    "type": "string",
                    "example": "PAUSED"
                },
                "topic": {
                    "type": "string",
                    "example": "my-notification-topic"
                }
            }
        },
...
        "types.UpdateWebhookEventStreamRequest": {
            "type": "object",
            "required": [
                "name",
                "url"
            ],
            "properties": {
                "chain": {
                    "type": "string",
                    "example": "mainnet"
                },
                "headers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "my-kafka-stream"
                },
                "status": {
                    "type": "string",
                    "example": "PAUSED"
                },
                "url": {
                    "type": "string",
                    "example": "https://my-event-steam-endpoint.com"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "JWTAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}